feat: delete-app command
